+++ source/examples/api/omni.isaac.core/visual_materials.py
@@ -13,9 +13,19 @@ import numpy as np
 simulation_app = SimulationApp({"headless": False})
 
 from omni.isaac.core import World
-from omni.isaac.core.objects import VisualCuboid, DynamicCuboid
+from omni.isaac.core.objects import VisualCuboid
+from omni.isaac.core.materials import OmniGlass
 
 my_world = World(stage_units_in_meters=0.01)
+
+glass = OmniGlass(
+    prim_path="/World/visual_cube_material",
+    name="glass_name",
+    ior=1.25,
+    depth=1.0,
+    thin_walled=True,
+    color=np.array([1, 0, 0]),
+)
 
 cube_1 = my_world.scene.add(
     VisualCuboid(
@@ -24,29 +34,12 @@ cube_1 = my_world.scene.add(
         position=np.array([0, 0, 0.5]) * 100,
         size=np.array([0.3, 0.3, 0.3]) * 100,
         color=np.array([255, 255, 255]),
+        visual_material=glass,
     )
 )
 
-cube_2 = my_world.scene.add(
-    DynamicCuboid(
-        prim_path="/new_cube_2",
-        name="cube_1",
-        position=np.array([0, 0, 1.0]) * 100,
-        size=np.array([0.6, 0.5, 0.2]) * 100,
-        color=np.array([255, 0, 0]),
-    )
-)
-
-cube_3 = my_world.scene.add(
-    DynamicCuboid(
-        prim_path="/new_cube_3",
-        name="cube_2",
-        position=np.array([0, 0, 3.0]) * 100,
-        size=np.array([0.1, 0.1, 0.1]) * 100,
-        color=np.array([0, 0, 255]),
-        linear_velocity=np.array([0, 0, 0.4]),
-    )
-)
+visual_material = cube_1.get_applied_visual_material()
+visual_material.set_color(np.array([1.0, 0.5, 0.0]))
 
 my_world.scene.add_ground_plane()
 
@@ -54,7 +47,5 @@ for i in range(5):
     my_world.reset()
     for i in range(1000):
         my_world.step(render=True)
-        print(cube_2.get_angular_velocity())
-        print(cube_2.get_world_pose())
 
 simulation_app.close()