+++ source/extensions/omni.isaac.core/omni/isaac/core/objects/capsule.py
@@ -28,7 +28,7 @@ class VisualCapsule(GeometryPrim):
         color: Optional[np.ndarray] = None,
         radius: float = 0.5,
         height: float = 0.5,
-        visual_material_path=None,
+        visual_material=None,
     ) -> None:
         """[summary]
 
@@ -57,13 +57,12 @@ class VisualCapsule(GeometryPrim):
         )
         VisualCapsule.set_radius(self, radius)
         VisualCapsule.set_height(self, height)
-        if visual_material_path is None:
-            if color is None:
-                color = np.array([0.5, 0.5, 0.5])
-            preview_surface = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
-        else:
-            preview_surface = PreviewSurface(prim_path=visual_material_path)
-        VisualCapsule.apply_visual_material(self, preview_surface)
+        if not self.is_visual_material_applied():
+            if visual_material is None:
+                if color is None:
+                    color = np.array([0.5, 0.5, 0.5])
+                visual_material = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
+            VisualCapsule.apply_visual_material(self, visual_material)
         return
 
     def set_radius(self, radius: float) -> None:
@@ -119,7 +118,7 @@ class DynamicCapsule(RigidPrim, GeometryPrim):
         radius: float = 0.5,
         height: float = 0.5,
         physics_material_path=None,
-        visual_material_path=None,
+        visual_material=None,
     ) -> None:
         """[summary]
 
@@ -173,13 +172,12 @@ class DynamicCapsule(RigidPrim, GeometryPrim):
         DynamicCapsule.set_radius(self, radius)
         DynamicCapsule.set_height(self, height)
         # create visual material
-        if visual_material_path is None:
-            if color is None:
-                color = np.array([0.5, 0.5, 0.5])
-            preview_surface = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
-        else:
-            preview_surface = PreviewSurface(prim_path=visual_material_path)
-        DynamicCapsule.apply_visual_material(self, preview_surface)
+        if not self.is_visual_material_applied():
+            if visual_material is None:
+                if color is None:
+                    color = np.array([0.5, 0.5, 0.5])
+                visual_material = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
+            DynamicCapsule.apply_visual_material(self, visual_material)
 
         if physics_material_path is None:
             my_physics_material = PhysicsMaterial(