+++ source/extensions/omni.isaac.samples/python/scripts/hello_world/hello_world.py
@@ -8,7 +8,7 @@
 
 from omni.isaac.samples.scripts.base_sample import BaseSample
 import numpy as np
-from omni.isaac.core.objects import DynamicCube, FixedCube
+from omni.isaac.core.objects import DynamicCuboid, FixedCuboid
 
 
 class HelloWorld(BaseSample):
@@ -20,20 +20,20 @@ class HelloWorld(BaseSample):
         world = self.get_world()
         world.scene.add_ground_plane()
         dynamic_cube = world.scene.add(
-            DynamicCube(
+            DynamicCuboid(
                 prim_path="/World/cube",
                 name="my_first_cube",
                 position=np.array([0, 0, 0.8]) * 100,
-                size=0.2 * 100,
+                size=np.array([0.2, 0.2, 0.2]) * 100,
                 color=np.array([1.0, 1.0, 1.0]),
             )
         )
         fixed_cube = world.scene.add(
-            FixedCube(
+            FixedCuboid(
                 prim_path="/World/fixed_cube",
                 name="my_first_fixed_cube",
                 position=np.array([0, 0, 0.5]) * 100,
-                size=0.2 * 100,
+                size=np.array([0.2, 0.2, 0.2]) * 100,
                 color=np.array([0, 0.5, 0]),
             )
         )